Updated build scripts to add webengine resources to final package. Updated README.md
For those who read git commit messages, development has been a bit slower due to legal issues concerning the name of the software. We expect that within the next week we will rename the software, and move to a more regular update schedule. Future goals include streamlining the installation & update process, continuing the replacement of the selenium code with playwright, and further improving performance for large numbers of nodes.