Add ability to select LQL Query result nodes on current canvas.
Add query conditions for the sum of the values of the primary fields of parents & children.

Add query conditions for total number of ancestors & successors.

Made it so 'Run Query' button is preselected, and that you can run multiple queries in the current context without needing to reset the wizard each time.